further ui comps reorg
